// src/constant/panel.type.ts
enum PanelType {
	BAKE = "bake",
	LIGHTS = "lights",
	LIGHTPORBES = "lightporbes",
	SKY = "sky",
	VIEWER = "viewer",
	GIF = "gif",
	IMAGE = "image",
	// 音频
	MUSIC = "music",
	MODEL = "model",
	// 环物图
	RING = "ring",
	TEXT = "text",
	VIDEO = "video",
	HOT = "hot",
	MATERIALS = "materials",
	OBJECTS = "objects",
	CAMERA = "camera",
	SETUP = "setup",
	// ANIMATION = 'animation',
	EFFECT = "effect",
	OUTLINE = "outline",
	POSITION = "position",
	POSITIONEDIY = "positionEdit",

	MATERIALEDIT = "materialEdit"
}
export default PanelType